requirements.txt
numpy>=1.20
soundfile>=0.10
librosa>=0.9
requests>=2.25
python-dotenv>=1.0
streamlit>=1.28